test: added proxy lookup behaviour unit tests refactor: pulled the extensive mock setup code out of lookup_verb_test.dart and into test_utils.dart so all of that mock setup (e.g. for the pol handshake simulation) can be reused by other verbs which use outbound connections
